Move InstallButton into core
