+++ tests/client/amo/components/TestAddonDetail.js
@@ -8,7 +8,7 @@ import {
 import AddonDetail, { allowedDescriptionTags }
   from 'amo/components/AddonDetail';
 import I18nProvider from 'core/i18n/Provider';
-import InstallButton from 'disco/components/InstallButton';
+import InstallButton from 'core/components/InstallButton';
 
 import { getFakeI18nInst } from 'tests/client/helpers';
