+++ tests/client/core/reducers/test_addons.js
@@ -1,5 +1,5 @@
 import addons from 'core/reducers/addons';
-import { API_THEME_TYPE, THEME_TYPE } from 'disco/constants';
+import { API_THEME_TYPE, THEME_TYPE } from 'core/constants';
 
 describe('addon reducer', () => {
   let originalState;
