+++ tests/client/helpers.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { createRenderer } from 'react-addons-test-utils';
-import { EXTENSION_TYPE } from 'disco/constants';
+import { EXTENSION_TYPE } from 'core/constants';
 import { sprintf } from 'jed';
 
 export function shallowRender(stuff) {
